refactor Transaction Review component
