+++ apps/mobile/components/transaction-review/TransactionReview.tsx
@@ -22,55 +22,31 @@
 
 import { PeriodFilterModal } from "@/components/modals/PeriodFilterModal";
 import { TypeFilterModal } from "@/components/modals/TypeFilterModal";
-import { useToast } from "@/components/ui/Toast";
 import { palette } from "@/constants/colors";
 import { useTheme } from "@/context/ThemeContext";
-import { useCategories } from "@/hooks/useCategories";
-import { useMarketRates } from "@/hooks/useMarketRates";
-import { getPeriodDateRange } from "@/hooks/usePeriodSummary";
-import type {
-  GroupingPeriod,
-  TransactionTypeFilter,
-} from "@/hooks/useTransactionsGrouping";
-import type { PendingAccount } from "@/services/pending-account-service";
-import {
-  type AccountMatch,
-  type AccountWithBankDetails,
-  fetchAccountsWithDetails,
-  matchTransactionsBatched,
-} from "@/services/sms-account-matcher";
-import { prepareSavePayload } from "@/services/sms-review-save-service";
-import { getCurrentUserId } from "@/services/supabase";
 import type { ReviewableTransaction } from "@astik/logic";
 import { Ionicons } from "@expo/vector-icons";
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import {
-  ActivityIndicator,
-  FlatList,
-  Text,
-  TouchableOpacity,
-  View,
-} from "react-native";
+import React, { useCallback } from "react";
+import { FlatList, Text, TouchableOpacity, View } from "react-native";
 import Animated, { FadeInDown } from "react-native-reanimated";
-import {
-  TransactionEditModal,
-  type TransactionEdits,
-} from "./TransactionEditModal";
+import { TransactionEditModal } from "./TransactionEditModal";
 import { TransactionItem } from "./TransactionItem";
-import { useCategoryLookup } from "@/context/CategoriesContext";
 import { TransactionFiltersBar } from "@/components/transactions/TransactionFiltersBar";
+import {
+  type ReviewListItem,
+  useTransactionReviewState,
+} from "./useTransactionReviewState";
+import {
+  getExpandedContent,
+  OriginalContentBlock,
+} from "./get-expanded-content";
+import { ReviewActionBar } from "./ReviewActionBar";
 
 // ---------------------------------------------------------------------------
 // Types
 // ---------------------------------------------------------------------------
 
-interface TransactionReviewProps {
+export interface TransactionReviewProps {
   /** All parsed transactions from the scan */
   readonly transactions: readonly ReviewableTransaction[];
   /** Called when user saves selected transactions with their account mappings */
@@ -85,105 +61,6 @@ interface TransactionReviewProps {
   readonly isSaving: boolean;
 }
 
-/** List item — either a date header or a transaction row */
-type ReviewListItem =
-  | { readonly kind: "header"; readonly date: string; readonly key: string }
-  | {
-      readonly kind: "transaction";
-      readonly originalIndex: number;
-      readonly tx: ReviewableTransaction;
-      readonly key: string;
-    };
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function groupByDate(
-  transactions: readonly ReviewableTransaction[],
-  originalTransactions: readonly ReviewableTransaction[]
-): readonly ReviewListItem[] {
-  const items: ReviewListItem[] = [];
-  let lastDate = "";
-
-  // Build a lookup from transaction reference → original index
-  const originalIndexMap = new Map<ReviewableTransaction, number>();
-  originalTransactions.forEach((tx, i) => originalIndexMap.set(tx, i));
-
-  // Transactions are assumed to be roughly sorted by date from the scan
-  const sorted = [...transactions].sort(
-    (a, b) => b.date.getTime() - a.date.getTime()
-  );
-
-  sorted.forEach((tx) => {
-    const dateKey = tx.date.toLocaleDateString("en-EG", {
-      weekday: "short",
-      day: "numeric",
-      month: "short",
-      year: "numeric",
-    });
-
-    if (dateKey !== lastDate) {
-      items.push({ kind: "header", date: dateKey, key: `h-${dateKey}` });
-      lastDate = dateKey;
-    }
-
-    const originalIndex = originalIndexMap.get(tx) ?? 0;
-    items.push({
-      kind: "transaction",
-      originalIndex,
-      tx,
-      key: `tx-${originalIndex}`,
-    });
-  });
-
-  return items;
-}
-
-/**
- * Apply period, type, and search filters on in-memory parsed transactions.
- * Returns a new array of filtered transactions.
- */
-function applyFilters(
-  transactions: readonly ReviewableTransaction[],
-  period: GroupingPeriod,
-  selectedTypes: readonly TransactionTypeFilter[],
-  searchQuery: string
-): readonly ReviewableTransaction[] {
-  let filtered = [...transactions];
-
-  // Period filter
-  if (period !== "all_time") {
-    const { startDate, endDate } = getPeriodDateRange(period);
-    filtered = filtered.filter((tx) => {
-      const time = tx.date.getTime();
-      return time >= startDate && time <= endDate;
-    });
-  }
-
-  // Type filter — map ReviewableTransaction type to TransactionTypeFilter
-  const includesAll = selectedTypes.includes("All");
-  if (!includesAll && selectedTypes.length > 0) {
-    filtered = filtered.filter((tx) => {
-      const txType = tx.type === "INCOME" ? "Income" : "Expense";
-      return selectedTypes.includes(txType);
-    });
-  }
-
-  // Search filter (uses generic originLabel instead of SMS-specific senderDisplayName)
-  if (searchQuery.trim()) {
-    const lower = searchQuery.trim().toLowerCase();
-    filtered = filtered.filter(
-      (tx) =>
-        tx.counterparty?.toLowerCase().includes(lower) ||
-        tx.originLabel.toLowerCase().includes(lower) ||
-        tx.amount.toString().includes(lower)
-    );
-  }
-
-  return filtered;
-}
-
 // ---------------------------------------------------------------------------
 // Component
 // ---------------------------------------------------------------------------
@@ -196,284 +73,7 @@ export function TransactionReview({
 }: TransactionReviewProps): React.JSX.Element {
   const { isDark } = useTheme();
 
-  // ── Filter state ──────────────────────────────────────────────────
-  const [period, setPeriod] = useState<GroupingPeriod>("all_time");
-  const [selectedTypes, setSelectedTypes] = useState<TransactionTypeFilter[]>([
-    "All",
-  ]);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [periodModalVisible, setPeriodModalVisible] = useState(false);
-  const [typeModalVisible, setTypeModalVisible] = useState(false);
-
-  // ── Selection state ─────────────────────────────────────────────────
-  const [selectedIndices, setSelectedIndices] = useState<ReadonlySet<number>>(
-    () => new Set(transactions.map((_, i) => i)) // All selected by default
-  );
-
-  const selectedIndicesRef = useRef(selectedIndices);
-  selectedIndicesRef.current = selectedIndices;
-
-  // ── Unified transaction overrides ─────────────────────────────────
-  // Stores all user edits (amount, counterparty, type, category, account)
-  // keyed by original transaction index.
-  const [transactionOverrides, setTransactionOverrides] = useState<
-    ReadonlyMap<number, TransactionEdits>
-  >(new Map());
-
-  // ── Account matching state ────────────────────────────────────────
-  const [accountMatches, setAccountMatches] = useState<
-    ReadonlyMap<number, AccountMatch>
-  >(new Map());
-  const [userAccounts, setUserAccounts] = useState<
-    readonly AccountWithBankDetails[]
-  >([]);
-
-  // ── Pending accounts (in-memory, created via "+ New" in edit modal) ──
-  const [pendingAccounts, setPendingAccounts] = useState<
-    readonly PendingAccount[]
-  >([]);
-
-  // ── Missing info flags (set on failed save validation) ──
-  const [invalidIndices, setInvalidIndices] = useState<ReadonlySet<number>>(
-    new Set()
-  );
-
-  const { showToast } = useToast();
-
-  const handleCreatePendingAccount = useCallback((account: PendingAccount) => {
-    setPendingAccounts((prev) => [...prev, account]);
-  }, []);
-
-  // ── Market rates for currency conversion ────────────────────────
-  const { latestRates } = useMarketRates();
-
-  // ── Inline edit modal state ──────────────────────────────────────
-  const [editModalIndex, setEditModalIndex] = useState<number | null>(null);
-
-  const { expenseCategories, incomeCategories } = useCategories();
-
-  const categoryMap = useCategoryLookup();
-
-  const batchSize = 20;
-
-  // Run batched account matching on mount (progressive rendering)
-  useEffect(() => {
-    let cancelled = false;
-    const run = async (): Promise<void> => {
-      try {
-        const userId = await getCurrentUserId();
-        if (!userId || cancelled) return;
-
-        // Fetch accounts first (also used for edit modal dropdown)
-        const accounts = await fetchAccountsWithDetails(userId);
-        if (!cancelled) {
-          setUserAccounts(accounts);
-        }
-
-        // Progressive batch matching (~20 txns/batch)
-        // Pass pre-fetched accounts to avoid duplicate DB query
-        await matchTransactionsBatched(
-          transactions,
-          userId,
-          batchSize,
-          (batchResults) => {
-            if (cancelled) return;
-            setAccountMatches((prev) => {
-              const next = new Map(prev);
-              for (const [idx, match] of batchResults) {
-                next.set(idx, match);
-              }
-              return next;
-            });
-          },
-          accounts
-        );
-      } catch (err: unknown) {
-        if (cancelled) return;
-        console.warn("[TransactionReview] Account matching failed:", err);
-        showToast({
-          type: "warning",
-          title: "Account Matching Failed",
-          message:
-            "Some transactions may not have an account assigned. You can assign them manually.",
-          duration: 4000,
-        });
-      }
-    };
-    // Inner catch handles all errors — no-op catch for the promise chain
-    run().catch(() => undefined);
-    return () => {
-      cancelled = true;
-    };
-  }, [transactions, showToast]);
-
-  // ── Derived data ──────────────────────────────────────────────────
-  const effectiveTransactions =
-    useMemo((): readonly ReviewableTransaction[] => {
-      return transactions.map((tx, i) => {
-        const overrides = transactionOverrides.get(i);
-        if (!overrides) return tx;
-        return {
-          ...tx,
-          amount: overrides.amount,
-          type: overrides.type,
-          categoryId: overrides.categoryId,
-          // Refresh display name when category changes so the UI label stays in sync.
-          categoryDisplayName:
-            categoryMap.get(overrides.categoryId)?.displayName ??
-            tx.categoryDisplayName,
-          // Only apply optional overrides when explicitly set by the user.
-          // Spreading undefined would erase the original transaction value.
-          ...(overrides.counterparty !== undefined && {
-            counterparty: overrides.counterparty,
-          }),
-        };
-      });
-    }, [transactions, transactionOverrides, categoryMap]);
-
-  // Apply filters before grouping
-  const filteredTransactions = useMemo(
-    () =>
-      applyFilters(effectiveTransactions, period, selectedTypes, searchQuery),
-    [effectiveTransactions, period, selectedTypes, searchQuery]
-  );
-
-  const listItems = useMemo(
-    () => groupByDate(filteredTransactions, effectiveTransactions),
-    [filteredTransactions, effectiveTransactions]
-  );
-
-  // Selection counts based on filtered view
-  const filteredOriginalIndices = useMemo(() => {
-    const indexMap = new Map<ReviewableTransaction, number>();
-    effectiveTransactions.forEach((tx, i) => indexMap.set(tx, i));
-    return filteredTransactions.map((tx) => indexMap.get(tx) ?? 0);
-  }, [filteredTransactions, effectiveTransactions]);
-
-  const allSelected =
-    filteredTransactions.length > 0 &&
-    filteredOriginalIndices.every((i) => selectedIndices.has(i));
-  const selectedCount = selectedIndices.size;
-
-  // ── Callbacks ─────────────────────────────────────────────────────
-
-  const handleToggleAll = useCallback(() => {
-    if (allSelected) {
-      setSelectedIndices(new Set());
-    } else {
-      setSelectedIndices(new Set(transactions.map((_, i) => i)));
-    }
-  }, [allSelected, transactions]);
-
-  const handleToggleItem = useCallback((index: number) => {
-    setSelectedIndices((prev) => {
-      const next = new Set(prev);
-      if (next.has(index)) {
-        next.delete(index);
-      } else {
-        next.add(index);
-      }
-      return next;
-    });
-  }, []);
-
-  const handleOpenEditModal = useCallback((index: number) => {
-    setEditModalIndex(index);
-  }, []);
-
-  const handleEditModalSave = useCallback(
-    (edits: TransactionEdits) => {
-      if (editModalIndex === null) return;
-
-      // Merge new edits into existing overrides for this index.
-      // Strip undefined values so optional fields the user didn't touch
-      // (e.g. counterparty) don't overwrite previously saved edits.
-      setTransactionOverrides((prev) => {
-        const next = new Map(prev);
-        const existing = next.get(editModalIndex);
-        const definedEdits = Object.fromEntries(
-          Object.entries(edits).filter(([, v]) => v !== undefined)
-        );
-        // Object.assign preserves the target type without needing a cast
-        const merged: TransactionEdits = Object.assign(
-          {},
-          existing,
-          definedEdits
-        );
-        next.set(editModalIndex, merged);
-        return next;
-      });
-
-      setEditModalIndex(null);
-    },
-    [editModalIndex]
-  );
-
-  const handleSave = useCallback(async (): Promise<void> => {
-    const userId = await getCurrentUserId();
-    if (!userId) {
-      showToast({
-        type: "error",
-        title: "Save Error",
-        message: "User not authenticated.",
-      });
-      return;
-    }
-
-    const result = await prepareSavePayload({
-      selectedIndices,
-      transactionOverrides,
-      accountMatches,
-      pendingAccounts,
-      effectiveTransactions,
-      userId,
-    });
-
-    if (!result.success) {
-      if (result.reason === "missing_accounts") {
-        setInvalidIndices(result.missingIndices);
-        showToast({
-          type: "warning",
-          title: "Missing Info",
-          message: result.message,
-          duration: 4000,
-        });
-      } else {
-        showToast({
-          type: "error",
-          title: "Account Creation Failed",
-          message: result.message,
-        });
-      }
-      return;
-    }
-
-    // Clear any previous flags
-    setInvalidIndices(new Set());
-
-    await onSave(
-      result.selected,
-      result.transactionAccountMap,
-      result.toAccountMap
-    );
-  }, [
-    effectiveTransactions,
-    selectedIndices,
-    accountMatches,
-    transactionOverrides,
-    pendingAccounts,
-    onSave,
-    showToast,
-  ]);
-
-  const handleTypeToggle = useCallback((type: TransactionTypeFilter) => {
-    setSelectedTypes((prev) => {
-      if (prev.includes(type)) {
-        return prev.filter((t) => t !== type);
-      }
-      return [...prev, type];
-    });
-  }, []);
+  const state = useTransactionReviewState({ transactions, onSave });
 
   // ── Render ────────────────────────────────────────────────────────
   const renderItem = useCallback(
@@ -490,49 +90,36 @@ export function TransactionReview({
       if (!tx) return null;
 
       const accountName =
-        transactionOverrides.get(item.originalIndex)?.accountName ??
-        accountMatches.get(item.originalIndex)?.accountName ??
+        state.transactionOverrides.get(item.originalIndex)?.accountName ??
+        state.accountMatches.get(item.originalIndex)?.accountName ??
         null;
+
+      const content = getExpandedContent(tx);
 
       return (
         <TransactionItem
           transaction={tx}
           index={item.originalIndex}
-          isSelected={selectedIndicesRef.current.has(item.originalIndex)}
+          isSelected={state.selectedIndicesRef.current.has(item.originalIndex)}
           accountName={accountName}
           expandedContent={
-            "rawSmsBody" in tx && tx.rawSmsBody ? (
-              <View className="bg-slate-900/60 rounded-xl p-3">
-                <Text className="text-xs text-slate-500 mb-1 font-medium">
-                  Original SMS
-                </Text>
-                <Text className="text-xs text-slate-400 leading-5">
-                  {tx.rawSmsBody as string}
-                </Text>
-              </View>
-            ) : "originalTranscript" in tx && tx.originalTranscript ? (
-              <View className="bg-slate-900/60 rounded-xl p-3">
-                <Text className="text-xs text-slate-500 mb-1 font-medium">
-                  Original Note
-                </Text>
-                <Text className="text-xs text-slate-400 leading-5">
-                  {tx.originalTranscript as string}
-                </Text>
-              </View>
+            content ? (
+              <OriginalContentBlock title={content.title} body={content.body} />
             ) : undefined
           }
-          onToggleSelect={handleToggleItem}
-          onPress={handleOpenEditModal}
-          hasMissingInfo={invalidIndices.has(item.originalIndex)}
+          onToggleSelect={state.handleToggleItem}
+          onPress={state.handleOpenEditModal}
+          hasMissingInfo={state.invalidIndices.has(item.originalIndex)}
         />
       );
     },
     [
-      accountMatches,
-      transactionOverrides,
-      invalidIndices,
-      handleToggleItem,
-      handleOpenEditModal,
+      state.accountMatches,
+      state.transactionOverrides,
+      state.invalidIndices,
+      state.handleToggleItem,
+      state.handleOpenEditModal,
+      state.selectedIndicesRef,
     ]
   );
 
@@ -542,13 +129,13 @@ export function TransactionReview({
     <View className="flex-1">
       {/* ── Filters & Search ────────────────────────────────────── */}
       <TransactionFiltersBar
-        period={period}
-        onPeriodPress={() => setPeriodModalVisible(true)}
-        selectedTypes={selectedTypes}
+        period={state.period}
+        onPeriodPress={() => state.setPeriodModalVisible(true)}
+        selectedTypes={state.selectedTypes}
         allTypesCount={2}
-        onTypePress={() => setTypeModalVisible(true)}
-        searchQuery={searchQuery}
-        onSearchChange={setSearchQuery}
+        onTypePress={() => state.setTypeModalVisible(true)}
+        searchQuery={state.searchQuery}
+        onSearchChange={state.setSearchQuery}
         searchPlaceholder="Search counterparty, sender..."
         containerClassName="px-5 pb-2"
       />
@@ -560,33 +147,35 @@ export function TransactionReview({
       >
         <Text className="text-sm text-slate-600 dark:text-slate-300">
           <Text className="font-bold text-slate-900 dark:text-white">
-            {filteredTransactions.length}
+            {state.filteredTransactions.length}
           </Text>{" "}
           found ·{" "}
           <Text className="font-bold text-nileGreen-600 dark:text-nileGreen-400">
-            {selectedCount}
+            {state.selectedCount}
           </Text>{" "}
           selected
         </Text>
 
         <TouchableOpacity
-          onPress={handleToggleAll}
+          onPress={state.handleToggleAll}
           className="flex-row items-center"
           activeOpacity={0.7}
         >
           <Ionicons
-            name={allSelected ? "checkbox" : "square-outline"}
+            name={state.allSelected ? "checkbox" : "square-outline"}
             size={18}
-            color={allSelected ? palette.nileGreen[400] : palette.slate[400]}
+            color={
+              state.allSelected ? palette.nileGreen[400] : palette.slate[400]
+            }
           />
           <Text className="text-xs text-slate-400 ml-1.5">
-            {allSelected ? "Deselect All" : "Select All"}
+            {state.allSelected ? "Deselect All" : "Select All"}
           </Text>
         </TouchableOpacity>
       </Animated.View>
 
       {/* ── Transaction list ────────────────────────────────────── */}
-      {filteredTransactions.length === 0 ? (
+      {state.filteredTransactions.length === 0 ? (
         <View className="flex-1 items-center justify-center px-6">
           <Ionicons
             name="search"
@@ -599,10 +188,10 @@ export function TransactionReview({
         </View>
       ) : (
         <FlatList
-          data={listItems}
+          data={state.listItems}
           renderItem={renderItem}
           keyExtractor={keyExtractor}
-          extraData={selectedIndices}
+          extraData={state.selectedIndices}
           contentContainerClassName="px-4 pb-32"
           showsVerticalScrollIndicator={false}
           removeClippedSubviews
@@ -612,93 +201,56 @@ export function TransactionReview({
       )}
 
       {/* ── Bottom action bar ───────────────────────────────────── */}
-      <Animated.View
-        entering={FadeInDown.delay(200)}
-        className="px-5 pb-8 pt-4 bg-white/95 dark:bg-background-dark border-t border-slate-200 dark:border-slate-800 flex-row gap-4 items-center"
-      >
-        {/* Discard All */}
-        <TouchableOpacity
-          onPress={onDiscard}
-          disabled={isSaving}
-          activeOpacity={0.85}
-          className=" w-[50%] flex-1 py-4 rounded-xl items-center bg-slate-100 dark:bg-slate-800"
-        >
-          <Text className="text-slate-500 dark:text-slate-400 text-sm font-semibold">
-            Discard All
-          </Text>
-        </TouchableOpacity>
-
-        {/* Save Selected */}
-        <TouchableOpacity
-          onPress={() => {
-            handleSave().catch((err: unknown) => {
-              const message = err instanceof Error ? err.message : String(err);
-              showToast({
-                type: "error",
-                title: "Save Error",
-                message,
-              });
-            });
-          }}
-          disabled={selectedCount === 0 || isSaving}
-          activeOpacity={0.85}
-          className={`w-[50%] flex-1 py-4 rounded-xl items-center justify-center ${
-            selectedCount === 0 || isSaving
-              ? "bg-slate-300 dark:bg-slate-700"
-              : "bg-nileGreen-600"
-          }`}
-        >
-          {isSaving ? (
-            <ActivityIndicator color="white" />
-          ) : (
-            <Text className="text-white text-base font-bold">
-              Save {selectedCount} Transaction{selectedCount !== 1 ? "s" : ""}
-            </Text>
-          )}
-        </TouchableOpacity>
-      </Animated.View>
+      <ReviewActionBar
+        selectedCount={state.selectedCount}
+        isSaving={isSaving}
+        onSave={state.handleSave}
+        onDiscard={onDiscard}
+      />
 
       {/* ── Filter modals ────────────────────────────────────────── */}
       <PeriodFilterModal
-        visible={periodModalVisible}
-        selectedPeriod={period}
-        onSelect={setPeriod}
-        onClose={() => setPeriodModalVisible(false)}
+        visible={state.periodModalVisible}
+        selectedPeriod={state.period}
+        onSelect={state.setPeriod}
+        onClose={() => state.setPeriodModalVisible(false)}
       />
 
       <TypeFilterModal
-        visible={typeModalVisible}
-        selectedTypes={selectedTypes}
-        onToggle={handleTypeToggle}
-        onClose={() => setTypeModalVisible(false)}
+        visible={state.typeModalVisible}
+        selectedTypes={state.selectedTypes}
+        onToggle={state.handleTypeToggle}
+        onClose={() => state.setTypeModalVisible(false)}
       />
 
       {/* ── Inline edit modal ──────────────────────────────────── */}
-      {editModalIndex !== null && effectiveTransactions[editModalIndex] && (
-        <TransactionEditModal
-          visible={editModalIndex !== null}
-          transaction={effectiveTransactions[editModalIndex]}
-          currentAccountName={
-            transactionOverrides.get(editModalIndex)?.accountName ??
-            accountMatches.get(editModalIndex)?.accountName ??
-            null
-          }
-          currentAccountId={
-            transactionOverrides.get(editModalIndex)?.accountId ??
-            accountMatches.get(editModalIndex)?.accountId ??
-            null
-          }
-          accounts={userAccounts}
-          categoryMap={categoryMap}
-          pendingAccounts={pendingAccounts}
-          latestRates={latestRates}
-          expenseCategories={expenseCategories}
-          incomeCategories={incomeCategories}
-          onSave={handleEditModalSave}
-          onCreatePendingAccount={handleCreatePendingAccount}
-          onClose={() => setEditModalIndex(null)}
-        />
-      )}
+      {state.editModalIndex !== null &&
+        state.effectiveTransactions[state.editModalIndex] && (
+          <TransactionEditModal
+            visible={state.editModalIndex !== null}
+            transaction={state.effectiveTransactions[state.editModalIndex]}
+            currentAccountName={
+              state.transactionOverrides.get(state.editModalIndex)
+                ?.accountName ??
+              state.accountMatches.get(state.editModalIndex)?.accountName ??
+              null
+            }
+            currentAccountId={
+              state.transactionOverrides.get(state.editModalIndex)?.accountId ??
+              state.accountMatches.get(state.editModalIndex)?.accountId ??
+              null
+            }
+            accounts={state.userAccounts}
+            categoryMap={state.categoryMap}
+            pendingAccounts={state.pendingAccounts}
+            latestRates={state.latestRates}
+            expenseCategories={state.expenseCategories}
+            incomeCategories={state.incomeCategories}
+            onSave={state.handleEditModalSave}
+            onCreatePendingAccount={state.handleCreatePendingAccount}
+            onClose={() => state.setEditModalIndex(null)}
+          />
+        )}
     </View>
   );
 }
